Finish assignment 7 tasks
In the Week 7 assignment, we extend this functionality by integrating third-party APIs to suggest meal ideas based on the items in the user's shopping list. This new feature will not only make our application more interactive and useful, but it will also give us a chance to practice working with external APIs.